# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  secret-service_internal_domain.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      team_id:
        type: string
      updated_at:
        type: string
    type: object
  secret-service_internal_domain.ProjectMember:
    properties:
      created_at:
        type: string
      id:
        type: string
      project_id:
        type: string
      role:
        $ref: '#/definitions/secret-service_internal_domain.ProjectRole'
      user_id:
        type: string
    type: object
  secret-service_internal_domain.ProjectRole:
    enum:
    - admin
    - manager
    - developer
    type: string
    x-enum-varnames:
    - ProjectRoleAdmin
    - ProjectRoleManager
    - ProjectRoleDeveloper
  secret-service_internal_domain.ProjectTeam:
    properties:
      assigned_at:
        type: string
      assigned_by:
        type: string
      project_id:
        type: string
      team_id:
        type: string
    type: object
...
  secret-service_internal_dto.AddProjectMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    type: object
...
  secret-service_internal_dto.AssignTeamRequest:
    properties:
      role:
        description: Role assigned to each team member in the project. Defaults to
          "developer".
        type: string
      team_id:
        type: string
    type: object
  secret-service_internal_dto.CreateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Centralized API key storage and management service for development
    teams.
  title: Secret Service API
  version: "1.0"
paths:
...
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List user projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project name and description
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret-service_internal_domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /projects/{projectID}:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret-service_internal_domain.Project'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
...
  /projects/{projectID}/members:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.ProjectMember'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: User ID and role (admin/manager/developer)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.AddProjectMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Add member to project
      tags:
      - projects
  /projects/{projectID}/members/{userID}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove member from project (admin only)
      tags:
      - projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.AddProjectMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Change project member role (admin only)
      tags:
      - projects
...
  /projects/{projectID}/teams:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.ProjectTeam'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List teams assigned to a project
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Team ID and optional role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.AssignTeamRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Assign team to project (bulk-adds all team members)
      tags:
      - projects
  /projects/{projectID}/teams/{teamID}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove team link from project
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"